pyproject.toml
[tool.poetry]
name = "lucidtask"
version = "0.1.0"
description = "Test task"
authors = ["mregorlite"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
fastapi = ">=0.115.7,<0.116.0"
uvicorn = ">=0.34.0,<0.35.0"
sqlalchemy = ">=2.0.37,<3.0.0"
alembic = ">=1.14.1,<2.0.0"
python-dotenv = "^1.0.1"
pydantic = "^2.10.6"
pydantic-settings = "^2.8.0"
email-validator = "^2.2.0"
python-jose = "^3.4.0"
passlib = "^1.7.4"
gunicorn = "^23.0.0"
bcrypt = "^4.2.1"
fastapi-cache2 = "^0.2.2"
redis = "^5.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
